Require only numpy 1.7 for the benefit of readthedocs
ReadTheDocs comes with pre-built packages for the basic scientific python
stack, but some of these packages are old (e.g., numpy is 1.7.1). The only way
to upgrade packages on readthedocs is to use a virtual environment and a
requirements.txt.

Unfortunately, this means we can't upgrade both numpy and pandas
simultaneously, because pandas may get built first and link against the wrong
version of numpy. We inadvertantly stumbled upon a work around to build the
"latest" docs by first installing numpy in the (cached) virtual environment,
and then later (in another commit), adding pandas to the requirements.txt
file.

However, this is a real hack and makes it impossible to maintain different
versions of the docs, such as for tagged releases. Accordingly, this commit
relaxes the numpy version requirement so we can use a version that readthedocs
already has installed. (We actually don't really need a newer version of numpy
for any current functionality in xray, although it's nice to have for support
for missing value functions like nanmean.)