Support "None" as a variable name and use it as a default
This makes the xray API a little more similar to pandas, which
makes heavy use of `name = None` for objects that can but don't always
have names like Series and Index.

It will be a particular useful option to have around when we add
a direct constructor for DataArray objects (#115). For now, arrays will
probably only end up being named `None` if they are the result of some
mathematical operation where the name could be ambiguous.