Dataset context manager and close() method
With this PR, it is possible to close the data store from which Dataset
was loaded via `ds.close()` or automatically when a dataset is used with
a context manager:

	with xray.open_dataset('data.nc') as ds:
		...

The ability to cleanly close files opened from disk is pretty essential -- we
probably should have had this a while ago. It should not be necessary to use
the low-level/unstable datastore API to get this functionality.

Implementation question: With this current implementation, calling
`ds.close()` on (and using a context manager with) a dataset not linked to any
file objects is a no-op. Should we raise an exception instead? Something like
`IOError('no file object to close')`?

CC @ToddSmall