Fix fillna with dict arguments
Fixes #379

Also reworks Dataset in-place arithmetic to no longer modify arrays
in-place. This lets us add automatic alignment.

Although the old operations were technically in-place, we had to make
defensive copies of everything to ensure we could roll back if we something
went wrong. This version is simpler code and more fully featured, so it
seems like a clear win.