moved keep_attrs kwarg from func to apply for DataArray
The funcs passed in to DataArrayGroupBy.apply did not respect the
keep_attrs argument, and seem to keep attributes by default. Now
DataArrayGroupBy.apply does the work of clearing result.attrs if
keep_attrs is False, and does nothing to the attributes otherwise.

This still needs to be reflected in the docstring.