keep_attrs for Dataset.resample and DataArray.resample (#829)
* initial addition of keep_attrs argument to resample, and tests

resample now has a keep_attrs argument that is passed on to reduction
operations. However, it is not being applied in all cases. Some tests
have been added that do not pass.

* moved keep_attrs kwarg from func to apply for DataArray

The funcs passed in to DataArrayGroupBy.apply did not respect the
keep_attrs argument, and seem to keep attributes by default. Now
DataArrayGroupBy.apply does the work of clearing result.attrs if
keep_attrs is False, and does nothing to the attributes otherwise.

This still needs to be reflected in the docstring.

* DataArrayGroupBy concat no longer sometimes copies metadata

DataArrayGroupBy._concat_shortcut copied metadata from itself to the
applied arrays, when it should not. Metadata copying should be, and is,
handled elsewhere.

With this fix, keep_attrs for resample methods is implemented and passes
the written tests. No other tests are broken.

* default for resample is now keep_attrs=False, tests for Dataset.resample

Tests have been added to make sure that keep_attrs applies to
Dataset attributes as well as the attributes of variables in that
Dataset.

resample now has keep_attrs=False as its default regardless of the
operation being completed (previously it was True for 'first' and
'last')

* cleaned up tests for keep_attrs option of resample

* added "what's new" notes for keep_attrs=False resampling option