Added convenience method for saving DataArray to netCDF file (#990)
* Added convenience method for saving DataArray to netCDF file

Added a simple function to DataArray that creates a dataset with one variable
called 'data' and then saves it to a netCDF file. All parameters are passed through
to to_netcdf().

Added an equivalent function called `open_dataarray` to be used to load from these files.

* Updated open_dataarray and da.to_netcdf to respect names & load any single var dataset

* Added tests for da.to_netcdf() and xr.open_dataarray()

* Implemented perfect round-tripping for to_netcdf and open_dataarray

Changed name that is used if da has no name
Stored old name in attrs so it can be replaced
Updated tests
Updated docs

* Updated Sphinx docs

* Added appropriate requires decorator for tests

* Added close() for DataArray, and moved closing code to BaseDataObject

When loading a DataArray, the _file_obj attribute is set to the same object
as the source Dataset's _file_obj. This can then be used to close the file
from the DataArray.

Tests have been updated to use this

* Switched to full func signature, used context mgr in tests

* Updated whats new to reference issue & attribute changes

* Moved repeated strings to constants & tidied up docstrings