Shorter repr for attributes (#1322)
* Shorter repr for attributes

NetCDF files often have tens of attributes, including multi-paragraph
summaries or the full modification history of the file.  It's great to
have this available in the .attrs, but we can truncate it substantially
in the repr!  Hopefully this will stop people writing `data.attrs = {}`
and discarding metadata in interactive workflows for the sake of cleaner
output.

* Add test for attribute repr