Fix missing config errors when running test suite (#1365)
* Fix missing config errors when running test suite

Fixes GH1358, GH1353

The approach I took here is to *not* package `conftest.py` with the xarray
distribution, but rather make the test suite robust to it's absence.

Also, as suggested by QuLogic in GH1336, switched from using a custom decorator
for `--run-slow` to just using `@pytest.mark.slow`.

Finally, I speed-up the test_dataset.test_rolling_reduce tests (from ~55 to ~17
seconds when bottleneck is not installed) and marked them as slow. It would be
nice to get them running in a few seconds or less, but we can save that for
later. On my laptop, the whole test suite now runs with `-k 'not slow'` in
36 seconds, which is not great, but acceptable.

CC pwolfram jhamman ghisvail

* fix typo netCDF4 -> has_netCDF4