Fixed https://github.com/pydata/xarray/issues/1434. There were two issues: 1 . dtype of the dimension that we concatenate along, was lost 2. dtype of the other dimensions (ones that are not concatenated but aligned) was lost
Fixing 1. required small refactoring of the "concat" classmethod in the variables.py -  by default we check if the dimensions can be concatenated using numpy.concat function and if this method fails we resort to the current method of concatenation using append fcn from pandas

As far as 2. it turns out the dtype of the dimensions that are not concatenated was lost in the "align" function in alignment.py - see my comment with # GH1434 tag. the solution was to store dtypes of the original dimensions and reapply those to the dimensions DataArrays after "align" function call