Add tests for pynio python3 support
Add tests for development version of pynio with python3 support. There
is currently no official release version of pynio with python3 support,
so we need to pull in the development version from the ncar conda
channel to use pynio in python 3 environments and to run the pynio tests
in the test suite. Also fixes a bug in the test initialization in which
the presence of the pynio library was tested by trying to import "pynio"
instead of "Nio" (the actual name of the library).