Fix py3 pynio backend (#1612)
* Update pynio backend for python 3

Update pynio backend to work with python 3. Previous version accessed
the iteritems() method of of the Nio variables object, which no longer
exists in the python 3 version. Instead, access the items() method here.

* Add tests for pynio python3 support

Add tests for development version of pynio with python3 support. There
is currently no official release version of pynio with python3 support,
so we need to pull in the development version from the ncar conda
channel to use pynio in python 3 environments and to run the pynio tests
in the test suite. Also fixes a bug in the test initialization in which
the presence of the pynio library was tested by trying to import "pynio"
instead of "Nio" (the actual name of the library).

* Add notes to bug fixes documenting changes.

Add notes to big fixes section in documentation to highlight changes.

* Fix requirements for py36-pynio-dev test

Fix requirements for py36-pynio-dev test to remove unnecessary option.

* Fix py36-pynio tests

Added tests were only specified in the allowed failures and needed to be
added to the main include block as well.