Make Indexer classes not inherit from tuple. (#1705)
* Make Indexer classes not inherit from tuple.

I'm not entirely sure this is a good idea. The advantage is that it ensures that
all our indexing code is entirely explicit: everything that reaches a backend
*must* be an ExplicitIndexer. The downside is that it removes a bit of
internal flexibility: we can't just use tuples in place of basic indexers
anymore. On the whole, I think this is probably worth it but I would appreciate
feedback.

* Add validation to ExplicitIndexer classes

* Fix pynio test failure

* Rename and add comments

* flake8

* Fix windows test failure

* typo

* leftover from debugging