Refactor xarray.conventions into VariableCoder
I would like to propose refactoring xarray.conventions to use an interface
based on VariableCoder objects with encode() and decode() methods.

The idea is make it easier to write new backends, by making decoding variables
according to CF conventions as simple as calling decode() on each coder in a
list of coders, with encoding defined by calling the same list of encoders in
the opposite order.

As a proof of concept, here I implement only a single Coder. In addition to
making use of xarray's existing lazy indexing behavior, I have written it so
that dask arrays are decoded using dask.

Eventually, we should port all the coders in xarray.conventions to this new
format. This is probably best saved for future PRs -- help would be appreciated!