Clarify our pull request template. (#1753)
* Clarify our pull request template.

I have two goals here:
1. I'd like to clarify for new contributors what we actually expect here.
2. I'd like to ensure smaller, self-contained changes, e.g., for major new
   features or refactoring, and make it clear that it's OK to merge parts of
   larger projects before they are entirely documented.

Large changes (more a few hundred lines of code) are hard to review and
collaborate on. As long as the code is tested (even if some are marked
xfail), we should try to get these changes in faster so we can iterate on them.
It's OK if documentation comes later, as long as we document everything before
it's released. It's also OK to copy code when refactoring and delete old code
later: that is often the easiest way to make large changes.

rabernat's recent zarr PR (GH1528) is probably a good example of something that
would have been good to have merged in smaller pieces :).

CC pydata/xarray

* separate tests added and tests passed in checklist