isin: better docs, support older numpy and use dask.array.isin.
I added a section to the indexing docs showing how to use `isin` with `where`
to do idnexing.

Adding support for older numpy turned out to be pretty easy, so I added a copy
of np.isin in npcompat.

This PR also makes use of dask.array.isin, so that we can support a dask array
properly as the second argument:
https://github.com/dask/dask/pull/3363