Ensure proper enable_netcdftimeindex option is used under lazy decoding
Prior to this, opening a dataset with enable_netcdftimeindex set to True
and then accessing one of its variables outside the context manager would
lead to it being decoded with the default enable_netcdftimeindex
(which is False).  This makes sure that lazy decoding takes into account
the context under which it was called.