doc: corrections of contributing.rst (#2266)
* doc: fix typos in contributing.rst

`` :issue:`1234` ``, which renders to `` GH1234 ``, is replaced with
``:issue:`1234```, which renders to :issue:`1234`.

* doc: fix style in contributing.rst

The formatting has been fixed to be more consistent. The lines longer
than 90–95 characters have been wrapped where appropriate.

* doc: fix path to built htmls in contributing.rst

* doc: recommend shorter first lines of commit messages

GitHub shows only the first 72 characters of the first line of
the commit message as the commit title. If the line is longer,
an ugly wrap takes place.