DOC: refresh "Why xarray" and shorten top-level description
This documentation revamp builds upon rabernat's rewrite in GH2430.

The main change is that the three paragraph description felt too long to me, so
I moved the background paragraph on multi-dimensional arrays into the next
section, on "Why xarray". I also ended up rewriting most of that page, and
made a few adjustments to the FAQ and related projects pages.