Type checking with mypy (#2655)
* Type checking with mypy

The rest of the scientific Python stack doesn't seem to support type
annotations yet, but that's OK -- we can use this incrementally in xarray when
it seems appropriate, and may check a few bugs. I'm especially excited to use
this for internal functions, where we don't always bother with full docstrings
(e.g., what is the type of the ``variables`` argument?).

This includes:
1. various minor fixes to ensure that "mypy xarray" passes.
2. adding "mypy xarray" to our lint check on Travis-CI.

For reference, see "Using mypy with an existing codebase":
https://mypy.readthedocs.io/en/stable/existing_code.html

Question: are we OK with (2)? This means Travis-CI will fail if your code
causes mypy to error.

* Lint fix

* DOC: document mypy, don't run it in travis

* document how to run mypy

* fix type annotation

* Pin pytest to avoid pytest-cov failure

see https://github.com/pytest-dev/pytest-cov/pull/253

* Revert pytest pinning

* Revert "Revert pytest pinning"

This reverts commit cd187a64eb97ca917476046a2f33c2a995a18680.

* Revert "Pin pytest to avoid pytest-cov failure"

This reverts commit 87ba452fad10f9f3715edae12abb1f28153ce99c.