Change function signature of to_stacked_array
Previously, this function was passed a list of dimensions which should be
stacked together. However, @benbovy found that the function failed when the
_non-stacked_ dimensions were not shared across all variables. Thus, it is
easier to specify the dimensions which should remain unchanged, rather than the
dimensions to be stacked.

The function to_stacked_array now takes an argument ''sample_dim'' which defines
these non-stacked dimensions. If these dims are not shared accross all
variables than an error is raised.