Fixed empty array data error
When using create_append_test_data we used np.arange(...), which was incidently also
the default value of the zarr array when fill_value is set to None. So appending to the data
with compute=False and then expecting an error when asserting the source and target to be the same
failed the tests. Using random data passes the tests