Only change under, over colors if they were set by user
When modifying a colormap, we only want to preserve the under and over
colors of the original colormap if they were explicitly set by the user.
In _build_discrete_cmap this makes no difference, as the new_cmap
returned by mpl.colors.from_levels_and_colors has the same minimum and
maximum colors as its input, so the default under and over colors would
not change anyway and could be copied regardless. However, for clarity
and in case the same pattern is needed in future elsewhere, it is nicer
to add a checks for: whether the under color is the same as cmap(0),
only setting under for new_cmap if it is not; and whether the over color
is the same as cmap(cmap.N - 1), only setting over for new_cmap if it is
not.