Add more attributes from the CF conventions.
"grid_mapping" and "bounds" are not the only attributes in CF that
contain names of variables that are not sensible to interpret as the
primary variables of analysis (XArray data variables) and that should
remain associated with the variable with the abstract.

This change will keep most of the referenced variables associated with
the referencing variables, but will also keep variables referenced by
other variables.