linear interp with NaNs in nd indexer (#4233)
* Added test for nd interpolation with nan

* Now ignoring NaNs in missing._localize

When interpolating with an nd indexer that contains NaN's, the code previously threw a KeyError from the missing._localize function. This commit fixes this by swapping `np.min` and `np.max` with `np.nanmin` and `np.nanmax`, ignoring any NaN values.

* Added `@requires_scipy` to test. Also updated what's new.

* Added numpy>=1.18 checks with `LooseVersion`

* Added checks for np.datetime64 type

This means the PR now also works for numpy < 1.18, as long as index is not with datetime

* Removed `raise ValueError` from previous commit

It seems that np.min/max works in place of nanmin/nanmax for datetime types for numpy < 1.18, see https://github.com/pydata/xarray/pull/3924/files

* Added datetime `NaT` test.

Also added a test for `Dataset` to `test_interpolate_nd_with_nan`, and "Missing values are skipped." to the dosctring of `interp` and `interp_like` methods of `DataArray` and `Dataset`.